+++ model/src/main/proto/spine_examples/chatspn/chat/deletion_commands.proto
@@ -39,7 +39,7 @@ import "spine_examples/chatspn/identifiers.proto";
 
 // Tells to delete the chat.
 message DeleteChat {
-    option (is).java_type = "io.spine.examples.chatspn.chat.DeleteChatMixin";
+    option (is).java_type = "DeleteChatMixin";
 
     // The ID of the chat deletion process.
     ChatDeletionId id = 1;
