+++ model/src/main/proto/spine_examples/chatspn/chat/deletion_rejections.proto
@@ -45,7 +45,7 @@ import "spine_examples/chatspn/identifiers.proto";
 // - the chat has already been marked as deleted.
 //
 message ChatCannotBeMarkedAsDeleted {
-    option (is).java_type = "io.spine.examples.chatspn.chat.ChatCannotBeMarkedAsDeletedMixin";
+    option (is).java_type = "ChatCannotBeMarkedAsDeletedMixin";
 
     // The ID of the chat that couldn't be marked as deleted.
     ChatId id = 1;
